refactor (db): widen IDs, add soft delete + audit columns, drop redundant fields
Schema cleanup across all 50+ migrations and matching Go models:

- SERIAL/INTEGER FKs -> BIGSERIAL/BIGINT everywhere (forward-compat for event tables)
- Add deleted_at (gorm.DeletedAt) + created_by/updated_by audit columns to all
  mutable tenant-scoped tables; partial unique indexes on name/identifier
  WHERE deleted_at IS NULL; FK constraints to users(user_id) ON DELETE SET NULL
  added in 026 for tables created before users
- Drop users.fullname (use Profile.first_name/last_name); profiles.email
  (use users.email); profiles.timezone/language (use user_settings.timezone/locale);
  user_settings.preferred_language (use locale, BCP-47); sms_templates.metadata (dead column)
- Scope permissions.name uniqueness to (tenant_id, api_id, name)
- Scope template/role/api/client/etc. name uniqueness per-tenant
- ip_restriction_rules.ip_address: VARCHAR(50) -> INET; type enum dedupe
  (drop whitelist/blacklist synonyms); same INET treatment on auth_events
  and security_settings_audit
- Standardize OAuth hash columns: sms_otps.otp_hash, user_backup_codes.code_hash
  to TEXT NOT NULL with index
- Drop dead Metadata field from sms_templates model

Service-layer compatibility helpers (internal/service/user_compat.go) bridge
the dropped columns at runtime via gorm:"-" transient fields:
- computeFullname(): derives display name from Profile (DisplayName -> FirstName+LastName)
- hydrateProfileTransients(): fills Profile.Email from User; Timezone/Language from UserSetting
- hydrateUserSettingTransients(): mirrors Locale -> PreferredLanguage

Read paths updated to use these helpers:
- toUserServiceDataResult, profile.GetByUUID/GetByUserUUID/GetAll,
  user_setting.toUserSettingServiceDataResult, oauth_userinfo handler
- userRepo.FindBySubAndClientID preloads default Profile so middleware-loaded
  users carry name info for OIDC userinfo

Write paths updated:
- user_setting.CreateOrUpdate writes to Locale (uses preferredLanguage arg as fallback)
- profile repo's Email filter now JOINs users (since profiles.email was dropped)

Misc fixes uncovered during testing:
- security_setting.UpdateMFAConfig passed "general" to updateConfig's switch
  which only knew "mfa" -> always returned "invalid config type". Fixed to "mfa".
- oauth_discovery_test fixture updated for the 3 OAuth grant types added in
  the recent CIBA/device_code/token-exchange feature work.

Also includes pending SMS login / OTP / backup codes / account service
scaffolding (migrations 051/052, models, repos, services, handlers, routes).

NOT done in this commit (deferred):
- Audit column auto-population from request context
- Auto-creation of default Profile on user create (handlers should call
  ensureDefaultProfile explicitly to persist Fullname)
- user_tokens migration 029, oauth_refresh_tokens migration 045, user_token
  model and repo (blocked from edits by local hook; apply same BIGSERIAL/BIGINT
  pattern manually)

go build ./..., go vet ./..., go test ./... all green.

+++ internal/database/migration/004_create_tenant_settings_table.go
@@ -6,13 +6,15 @@ import (
 
 // CreateTenantSettingsTable creates the tenant_settings table for tenant-level
 // operational configuration (rate limits, audit, maintenance, feature flags).
+// This is a 1:1 config row per tenant — no soft delete or audit columns since
+// the row's lifecycle is bound to the parent tenant via ON DELETE CASCADE.
 func CreateTenantSettingsTable(db *gorm.DB) error {
 	sql := `
 -- CREATE TABLE
 CREATE TABLE IF NOT EXISTS tenant_settings (
-    tenant_setting_id   SERIAL PRIMARY KEY,
+    tenant_setting_id   BIGSERIAL PRIMARY KEY,
     tenant_setting_uuid UUID NOT NULL UNIQUE,
-    tenant_id           INTEGER NOT NULL,
+    tenant_id           BIGINT NOT NULL,
     rate_limit_config   JSONB DEFAULT '{}',
     audit_config        JSONB DEFAULT '{}',
     maintenance_config  JSONB DEFAULT '{}',
